requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.7
whitenoise>=6.5.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
